Module tc375_bsp::leds_and_buttons

This module contains structures handling the leds and button of Aurix TC375 Lite Kit.

The board has two leds wired into Pin5 and Pin6 of Port_00, and one button wired into Pin7 of Port_00. This module provides separate structures for the two led and one structure for handling the one button.

Examples

use mod tx375-bsp::leds_and_buttons::{Led1, Led2, Button1};
let mut led1 = Led1::new();
let button1 = Button1::new();
if button1.is_pressed() {
    led1.toggle();
}

Structs

Button1

Handles button Button1 wired to Pin7 of Port_00.

Led1

Handles led Led1 wired to Pin5 of Port_00.

Led2

Handles led Led2 wired to Pin6 of Port_00.